pyproject.toml
[tool.poetry]
name = "capyau-bot"
version = "0.1.0"
description = ""
authors = ["Paulo Matias <matias@ufscar.br>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = {extras = ["job-queue"], version = "^20.5"}
requests = "^2.31.0"
gdal = "^3.7.2"
haversine = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
